issue-2 add injector class. moves routes into their own folder.
