+++ api/manager.js
@@ -1,0 +1,1 @@
+let injector = new (require('./injector/Injector.js'))();
